+++ src/design/plone/contenttypes/behaviors/argomenti.py
@@ -30,6 +30,23 @@ class IArgomentiSchema(model.Schema):
         required=False,
         default=[],
     )
+    correlato_in_evidenza = RelationList(
+        title=_(
+            "correlato_in_evidenza_label", default="Correlato in evidenza"
+        ),
+        description=_(
+            "correlato_in_evidenza_help",
+            default="Seleziona un correlato da mettere in evidenza."
+            " contenuto.",
+        ),
+        value_type=RelationChoice(
+            title=_(u"Correlato in evidenza"),
+            vocabulary="plone.app.vocabularies.Catalog",
+        ),
+        required=False,
+        default=[],
+    )
+
     form.widget(
         "tassonomia_argomenti",
         RelatedItemsFieldWidget,
@@ -39,6 +56,12 @@ class IArgomentiSchema(model.Schema):
             "selectableTypes": ["Pagina Argomento"],
         },
     )
+    form.widget(
+        "correlato_in_evidenza",
+        RelatedItemsFieldWidget,
+        vocabulary="plone.app.vocabularies.Catalog",
+        pattern_options={"maximumSelectionSize": 1},
+    )
 
     dexteritytextindexer.searchable("tassonomia_argomenti")
 
@@ -47,11 +70,24 @@ class IArgomentiSchema(model.Schema):
 class IArgomenti(IArgomentiSchema):
     """ """
 
+    model.fieldset(
+        "correlati",
+        label=_("correlati_label", default="Contenuti collegati"),
+        fields=["correlato_in_evidenza"],
+    )
+    form.order_after(correlato_in_evidenza="IRelatedItems.relatedItems")
+
 
 @provider(IFormFieldProvider)
 class IArgomentiDocumento(IArgomentiSchema):
     """ """
 
+    model.fieldset(
+        "correlati",
+        label=_("correlati_label", default="Contenuti collegati"),
+        fields=["correlato_in_evidenza"],
+    )
+    form.order_after(correlato_in_evidenza="IRelatedItems.relatedItems")
     form.order_after(tassonomia_argomenti="IDublinCore.title")
 
 
@@ -62,9 +98,10 @@ class IArgomentiBando(IArgomentiSchema):
     model.fieldset(
         "correlati",
         label=_("correlati_label", default="Contenuti collegati"),
-        fields=["tassonomia_argomenti"],
+        fields=["tassonomia_argomenti", "correlato_in_evidenza"],
     )
     form.order_before(tassonomia_argomenti="IRelatedItems.relatedItems")
+    form.order_after(correlato_in_evidenza="IRelatedItems.relatedItems")
 
 
 @implementer(IArgomenti)