+++ CHANGES.rst
@@ -2,7 +2,22 @@ Changelog
 =========
 
 
-1.0a1 (unreleased)
+1.0.2 (unreleased)
+------------------
+
+- Nothing changed yet.
+
+
+1.0.1 (2020-12-14)
+------------------
+
+- Add `immagine_testata` new field in *design.plone.contenttypes.behavior.info_testata* behavior.
+  [cekk]
+- Add `correlato_in_evidenza` new field in *design.plone.contenttypes.behavior.argomenti* behavior.
+  [cekk]
+
+
+1.0.0 (2020-12-07)
 ------------------
 
 - Initial release.